src/settings.py
from __future__ import annotations

from pathlib import Path
from typing import Any

import yaml

ROOT = Path(__file__).resolve().parents[1]
DEFAULT_DOSE_RULES_PATH = ROOT / "config" / "dose_rules.yaml"


def load_dose_rules(path: Path = DEFAULT_DOSE_RULES_PATH) -> dict[str, Any]:
    if not path.exists():
        return {}
    with path.open("r", encoding="utf-8") as f:
        return yaml.safe_load(f) or {}


def save_dose_rules(data: dict[str, Any], path: Path = DEFAULT_DOSE_RULES_PATH) -> None:
    path.parent.mkdir(parents=True, exist_ok=True)
    with path.open("w", encoding="utf-8") as f:
        yaml.safe_dump(data, f, allow_unicode=True, sort_keys=False)
