pyproject.toml
[project]
name = "dialysis-cdss"
version = "0.1.0"
description = "透析中心月藥物調整建議與查房工作流原型"
requires-python = ">=3.11"
dependencies = [
  "anthropic>=0.49",
  "openpyxl>=3.1",
  "pandas>=2.2",
  "pyyaml>=6.0",
  "requests>=2.32",
  "streamlit>=1.32",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
